pyproject.toml
[tool.poetry]
name = "street-fighter-ii-ai"
version = "0.1.0"
description = ""
authors = ["dthelegend <wampamba.daudi@gmail.com>"]
readme = "README.md"
packages = [{include = "street_fighter_ii_ai"}]

[tool.poetry.scripts]
main = "street_fighter_ii_ai.main:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
gymnasium = "^0.28.1"
cmake = "^3.26.1"
gym-retro = {git = "https://github.com/MatPoliquin/stable-retro.git"}
nvidia-cudnn-cu11 = "8.6.0.163"
tensorflow = {version = "^2.12.0", extras = ["gpu"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
